目录删除之前先做判断  如果含有主题模型  就不给删除

+++ core/backend/src/main/java/io/dataease/controller/dataset/DataSetGroupController.java
@@ -9,6 +9,7 @@ import io.dataease.commons.constants.SysLogConstants;
 import io.dataease.commons.utils.AuthUtils;
 import io.dataease.commons.utils.DeLogUtils;
 import io.dataease.controller.datamodel.enums.DatamodelEnum;
+import io.dataease.controller.request.authModel.VAuthModelRequest;
 import io.dataease.controller.request.dataset.DataSetGroupRequest;
 import io.dataease.dto.SysLogDTO;
 import io.dataease.dto.authModel.VAuthModelDTO;
@@ -21,6 +22,7 @@ import io.dataease.service.dataset.DataSetGroupService;
 import io.dataease.service.kettle.KettleService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+import org.apache.commons.collections4.CollectionUtils;
 import org.apache.shiro.authz.annotation.Logical;
 import org.springframework.web.bind.annotation.*;
 import springfox.documentation.annotations.ApiIgnore;
@@ -75,11 +77,49 @@ public class DataSetGroupController {
     @ApiOperation("删除")
     @PostMapping("/delete/{id}")
     public void delete(@PathVariable String id) throws Exception {
-        CacheUtils.remove(modelCacheEnum.modeltree.getValue(), AuthUtils.getUser().getUserId());
-        DatasetGroup datasetGroup = dataSetGroupService.getScene(id);
-        SysLogDTO sysLogDTO = DeLogUtils.buildLog(SysLogConstants.OPERATE_TYPE.DELETE, SysLogConstants.SOURCE_TYPE.DATASET, id, datasetGroup.getPid(), null, null);
-        dataSetGroupService.delete(id);
-        DeLogUtils.save(sysLogDTO);
+        //删除前 先判断是否含有主题模型
+        VAuthModelRequest request = new VAuthModelRequest();
+        request.setModelType("dataset");
+        List<VAuthModelDTO> vAuthModelDTOS = vAuthModelService.queryAuthModel(request);
+        List<VAuthModelDTO> currentTree = getsTheCurrentTree(id, vAuthModelDTOS);
+        Boolean aBoolean = checkTreeDirType(currentTree);
+        if (aBoolean){
+            throw new Exception("该目录下含有主题模型，请先删除主题模型再删除目录");
+        } else {
+            CacheUtils.remove(modelCacheEnum.modeltree.getValue(), AuthUtils.getUser().getUserId());
+            DatasetGroup datasetGroup = dataSetGroupService.getScene(id);
+            SysLogDTO sysLogDTO = DeLogUtils.buildLog(SysLogConstants.OPERATE_TYPE.DELETE, SysLogConstants.SOURCE_TYPE.DATASET, id, datasetGroup.getPid(), null, null);
+            dataSetGroupService.delete(id);
+            DeLogUtils.save(sysLogDTO);
+        }
+    }
+    private Boolean checkTreeDirType(List<VAuthModelDTO> vAuthModelDTOS) {
+        if (CollectionUtils.isEmpty(vAuthModelDTOS)){
+            return false;
+        }
+        for (VAuthModelDTO vAuthModelDTO : vAuthModelDTOS) {
+            if (vAuthModelDTO.getDirType() == DatamodelEnum.MODEL_DIR.getValue()){
+                return true;
+            } else {
+                checkTreeDirType(vAuthModelDTO.getChildren());
+            }
+        }
+        return false;
+    }
+
+    private List<VAuthModelDTO> getsTheCurrentTree(String id, List<VAuthModelDTO> vAuthModelDTOS) {
+        if (CollectionUtils.isEmpty(vAuthModelDTOS)){
+            return null;
+        }
+        for (VAuthModelDTO vAuthModelDTO : vAuthModelDTOS) {
+            List<VAuthModelDTO> children = vAuthModelDTO.getChildren();
+            if (id.equals(vAuthModelDTO.getId())){
+                return children;
+            } else {
+                return getsTheCurrentTree(id,children);
+            }
+        }
+        return null;
     }
 
     @ApiIgnore
